splitup raster_renderer
